non-static factory (#261)
This PR refactors the factory design. The redesigned API works pretty much in the same way, but multiple factories can co-exist and encapsulate a "context" (as defined by the [Settings] section in the respective input file). The factory can be safely destructed once the model (or any needed object) is retrieved. The retrieved objects are shared pointers.

Some additional changes included in this PR:
- default integer dtype is no longer part of the global setting -- it is now treated just like the default floating point dtype outside the neml2 scope
- machine precision, tolerance, and the "tighter" tolerance are removed from the global setting
- machine precision is defined as `std::numeric_limits<>::epsilon()` for float and double. For half precision (i.e. `float16_t`) an approximation is given.
- `OptionsCollection` is renamed to `InputFile` which seems to be more intuitive
- `Factory::get_object_ptr` is renamed to `Factory::get_object` which always returns a shared pointer
- `Real` is removed from neml2, since it's not configurable there's no point in defining this additional layer of aliasing. `Real` is replaced with `double` wherever appropriate.
- `neml2::CScalar` is added as an alias of `c10:Scalar` which is a type-erased number type for integer, double, complex, etc. Some tensor operations which used to be hard-coded to take `Real` has been changed to take `neml2::CScalar`.

+++ include/neml2/tensors/MillerIndex.h
@@ -43,9 +43,9 @@ public:
   Scalar operator()(Size i) const;
 
   /// Most likely construction method -- make from three ints
-  static MillerIndex fill(Integer a,
-                          Integer b,
-                          Integer c,
+  static MillerIndex fill(int64_t a,
+                          int64_t b,
+                          int64_t c,
                           const TensorOptions & options = default_integer_tensor_options());
 
   /// Reduce to the greatest common demoninator
